bot-requirements.txt
discord.py[voice]
python-dotenv
sqlalchemy
mysql-connector-python
apscheduler